requirements.txt
click==6.6
GitPython==2.0.2
requests-ntlm==0.3.0
python-dateutil==2.5.3
tabulate==0.7.5
